… provider/testdata/benchmarks/bigquery-datasetaccess-3-bench.txt
… provider/testdata/benchmarks/bigquery-routine-1-bench.txt
… provider/testdata/benchmarks/bigquery-routine-2-bench.txt
… provider/testdata/benchmarks/bigquery-table-bench.txt
… provider/testdata/benchmarks/cloudfunctions-function-bench.txt
… provider/testdata/benchmarks/cloudrunv2-job-6-bench.txt
… provider/testdata/benchmarks/cloudrunv2-service-4-bench.txt
… provider/testdata/benchmarks/cloudrunv2-service-5-bench.txt
… provider/testdata/benchmarks/cloudrunv2-service-7-bench.txt
… provider/testdata/benchmarks/compute-disk-1-bench.txt
… provider/testdata/benchmarks/compute-disk-3-bench.txt
… provider/testdata/benchmarks/compute-firewall-bench.txt
… provider/testdata/benchmarks/dns-recordset-bench.txt
… provider/testdata/benchmarks/monitoring-alertpolicy-1-bench.txt
… provider/testdata/benchmarks/pubsub-subscription-bench.txt
… provider/testdata/benchmarks/pubsub-topic-bench.txt
… provider/testdata/benchmarks/secretmanager-secret-bench.txt
… provider/testdata/benchmarks/serviceaccount-account-bench.txt
… provider/testdata/benchmarks/sql-user-bench.txt
… provider/testdata/benchmarks/storage-bucket-bench.txt
… provider/testdata/benchmarks/storage-bucketobject-bench.txt