## Minimum viable benchmarking for providers.



This introduces a test framework for recording benchmarks by collecting trace spans from ProgramTest. After collecting spans, we filter to spans that are most relevant to the provider itself and record into a benchmark file format in order to support comparing the results with [benchstat](https://pkg.go.dev/golang.org/x/perf/cmd/benchstat)

### example benchstat of master vs v5 release candidate

```
                                                                                       │    v4.txt    │                v5.txt                     │
                                                                                       │    sec/op    │    sec/op           vs base               │
/dns-recordset/preview-initial/ResourceProvider/GetPluginInfo/                           574.2µ ± 10%    602.5µ ±  2%  +4.93% (p=0.001 n=180+120)
/dns-recordset/preview-initial/newPlugin/gcp                                             209.7m ±  0%    210.5m ±  1%  +0.37% (p=0.007 n=60+40)
/dns-recordset/preview-initial/ResourceProvider/CheckConfig/pulumi:providers:gcp         249.6m ±  2%    251.8m ±  3%       ~ (p=0.135 n=60+40)
/dns-recordset/preview-initial/ResourceProvider/Configure/                               161.5m ±  2%    164.6m ±  2%       ~ (p=0.171 n=60+40)
/dns-recordset/preview-initial/ResourceProvider/Check/gcp:dns/managedZone:ManagedZone    702.2µ ± 19%    791.7µ ± 10%       ~ (p=0.056 n=30+20)
/dns-recordset/preview-initial/ResourceProvider/Create/gcp:dns/managedZone:ManagedZone   766.7µ ± 19%    867.6µ ± 13%       ~ (p=0.151 n=30+20)
/dns-recordset/preview-initial/ResourceProvider/Check/gcp:compute/instance:Instance      928.3µ ± 13%   1016.6µ ± 10%  +9.51% (p=0.049 n=30+20)
/dns-recordset/preview-initial/ResourceProvider/Create/gcp:compute/instance:Instance     3.475m ± 19%    3.592m ±  7%       ~ (p=0.064 n=30+20)
/dns-recordset/preview-initial/ResourceProvider/Check/gcp:dns/recordSet:RecordSet        1.438m ± 19%    1.352m ± 29%       ~ (p=0.976 n=30+20)
/dns-recordset/preview-initial/ResourceProvider/Create/gcp:dns/recordSet:RecordSet       1.056m ± 10%    1.075m ± 27%       ~ (p=0.656 n=30+20)
/dns-recordset/preview-initial/ResourceProvider/GetSchema/                               162.3m ±  1%    164.5m ±  1%  +1.36% (p=0.002 n=90+60)

```

I've organized the code into three parts:

1) traceutils.go contains functions and interfaces for working with traces and spans in their various formats. I expect we'll want to move most of this into a library, probably in pulumi-trace-tool (and add tests for it when we do)

2) benchmarks.go is the meat of the benchmark testing; it's patterned off of  the test config+runners in providertest, and can hopefully migrate there.

3) performance_test.go and provider_test.go are examples using the benchmark test on test programs. 

I've also checked in the recorded benchmarks from the tests in performance_test.go as a baseline and as examples of what the results look like.

Resolves https://github.com/pulumi/home/issues/3167