@mjeffryes 2023-10-28
FYI @VenelinMartinov This is what I ended up writing to collect benchmarks