@mjeffryes 2023-11-08
Yea, I'm going to land this and we can work on factoring the mess out from GCP and making it reusable in the coming weeks. Since we're so close on the major version release, I might hold off until we land that branch, just to avoid adding merge conflicts for guin.